credit-risk-webapp/backend/requirements.txt
# FastAPI và ASGI Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Machine Learning
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0

# AI và NLP
google-genai>=0.2.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.0.0

# Data Processing
openpyxl>=3.1.0

# Word Report Generation
python-docx>=0.8.11
Pillow>=10.0.0

# RSS Feed
feedparser>=6.0.0
python-dateutil>=2.8.0

# CORS
fastapi-cors>=0.0.6

# Environment
python-dotenv>=1.0.0
